Utilizing orbit-db-benchmark-runner
moving benchmarks to legacy folder

removing runner folder

npm install orbit-db-benchmark-runner

new benchmarks

adding benchmark to CI

adding nyc to benchmark running

Making the benchmarks more CI-friendly
