+++ benchmarks/from-entry-hash.js
@@ -1,28 +1,27 @@
-const startIPFS = require('./utils/start-ipfs')
+const { startIpfs, stopIpfs, config } = require('orbit-db-test-utils')
 const createLog = require('./utils/create-log')
-const Log = require('../../../src/log')
+const Log = require('../src/log')
 
 const base = {
   prepare: async function () {
-    const { ipfs, repo } = await startIPFS('./ipfs-log-benchmarks/fromEntryHash/ipfs')
-
-    const { log, access, identity } = await createLog(ipfs, 'A')
+    const ipfsd = await startIpfs('js-ipfs', config)
+    const { log, access, identity } = await createLog(ipfsd.api, 'A')
     const refCount = 64
     process.stdout.clearLine()
     for (let i = 1; i < this.count + 1; i++) {
       process.stdout.write(`\r${this.name} / Preparing / Writing: ${i}/${this.count}`)
       await log.append('hello' + i, refCount)
     }
-    return { ipfs, repo, log, access, identity }
+    return { ipfsd, log, access, identity }
   },
-  cycle: async function ({ log, ipfs, access, identity }) {
-    await Log.fromEntryHash(ipfs, identity, log.heads.map(e => e.hash), {
+  cycle: async function ({ log, ipfsd, access, identity }) {
+    await Log.fromEntryHash(ipfsd.api, identity, log.heads.map(e => e.hash), {
       access,
       logId: log._id
     })
   },
-  teardown: async function ({ repo }) {
-    await repo.close()
+  teardown: async function ({ ipfsd }) {
+    await stopIpfs(ipfsd)
   }
 }
 
@@ -38,7 +37,7 @@ const stress = {
   }
 }
 
-const counts = [1, 100, 1000, 10000]
+const counts = [1, 100, 1000]
 const benchmarks = []
 for (const count of counts) {
   const c = { count }
