+++ benchmarks/get.js
@@ -1,24 +1,25 @@
-const startIPFS = require('./utils/start-ipfs')
+const { startIpfs, stopIpfs, config } = require('orbit-db-test-utils')
 const createLog = require('./utils/create-log')
 
 const base = {
   prepare: async function () {
-    const { ipfs, repo } = await startIPFS('./ipfs-log-benchmarks/ipfs')
-    const { log } = await createLog(ipfs, 'A')
+    const ipfsd = await startIpfs('js-ipfs', config)
+    const { log } = await createLog(ipfsd.api, 'A')
 
     process.stdout.clearLine()
+    let entry
     for (let i = 1; i < this.count + 1; i++) {
       process.stdout.write(`\r${this.name} / Preparing / Writing: ${i}/${this.count}`)
-      await log.append(`Hello World: ${i}`)
+      entry = await log.append(`Hello World: ${i}`)
     }
 
-    return { log, repo }
+    return { log, ipfsd, entry }
   },
-  cycle: async function ({ log }) {
-    return log.tails
+  cycle: async function ({ log, entry }) {
+    return log.get(entry.hash)
   },
-  teardown: async function ({ repo }) {
-    await repo.close()
+  teardown: async function ({ ipfsd }) {
+    await stopIpfs(ipfsd)
   }
 }
 
@@ -34,12 +35,12 @@ const stress = {
   }
 }
 
-const counts = [1, 100, 1000, 5000]
+const counts = [1, 100, 1000]
 const benchmarks = []
 for (const count of counts) {
   const c = { count }
-  benchmarks.push({ name: `tails-${count}-baseline`, ...base, ...c, ...baseline })
-  benchmarks.push({ name: `tails-${count}-stress`, ...base, ...c, ...stress })
+  benchmarks.push({ name: `get-${count}-baseline`, ...base, ...c, ...baseline })
+  benchmarks.push({ name: `get-${count}-stress`, ...base, ...c, ...stress })
 }
 
 module.exports = benchmarks
