+++ benchmarks/legacy/benchmark-append.js
