+++ benchmarks/legacy/benchmark-from-entry-hash.js
