+++ benchmarks/legacy/benchmark-join2.js
