+++ benchmarks/legacy/browser/benchmark-append-signed.html
