+++ benchmarks/legacy/browser/benchmark-join-signed.html
