+++ package.json
@@ -48,6 +48,7 @@
     "mocha": "~8.1.3",
     "mocha-headless-chrome": "^3.1.0",
     "nyc": "^15.0.1",
+    "orbit-db-benchmark-runner": "^1.0.3",
     "orbit-db-storage-adapter": "^0.5.3",
     "orbit-db-test-utils": "~0.11.0",
     "rimraf": "~3.0.2",
@@ -84,8 +85,8 @@
     "build:dist": "webpack --config webpack.config.js --display-modules --sort-modules-by size --mode production",
     "build:es5": "babel src --out-dir ./lib/es5/ --presets @babel/preset-env --plugins @babel/plugin-transform-runtime",
     "lint": "standard",
-    "benchmark": "node --expose-gc benchmarks/runner/index.js -r --grep baseline",
-    "benchmark:stress": "node --expose-gc benchmarks/runner/index.js -r --grep stress",
+    "benchmark": "nyc benchmark-runner -r -b --baselineLimit 1000",
+    "benchmark:stress": "benchmark-runner -r --grep stress",
     "examples": "http-server examples/browser"
   },
   "localMaintainers": [
